main.ts moved to src
