+++ manifest.json
@@ -6,5 +6,5 @@
   "author": "Manic Chuang",
   "authorUrl": "https://github.com/manic/obsidian-wrap-with-shortcuts",
   "js": "main.js",
-  "version": "1.3.0"
+  "version": "1.3.1"
 }
