+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "obsidian-wrap-with-shortcuts",
-  "version": "1.3.0",
+  "version": "1.3.1",
   "description": "Wrap selected text in custom tags with shortcuts.",
   "main": "main.js",
   "scripts": {
