+++ versions.json
@@ -1,4 +1,5 @@
 {
+  "1.3.1": "1.1.0",
   "1.3.0": "1.1.0",
   "1.2.0": "0.15.0",
   "1.0.1": "0.9.12",
